Expose the little capacity cost as a parameter
If a pipe or a supply has a capacity-related cost of zero it has no
contribution to the objective and contributes to constraints from
above, so the MIP solver will reasonably enough set it to whatever its
upper bound is.

This term makes this not happen, by adding a very small penalty to
making a giant pipe or supply even if there is no financial cost to it
in the inputs.

In this commit I expose it to the outside so I can write tests in
which it is zero, because what I'm testing is the interaction between
other costs that precisely balance.